fix: cut search results state out of app
